+++ HISTORY.rst
@@ -3,6 +3,13 @@
 History
 -------
 
+0.0.6 (2019-02-27)
+++++++++++++++++++
+
+* added langid to install_requires
+* new function `recogito.recogito_dump_to_spacy_ner` added
+* minor things
+
 0.0.3 (2019-02-25)
 ++++++++++++++++++